Refactor BarcodeScanner module: - Removed unused properties and methods from BarcodeScannerStateModel. - Consolidated food search functionality into a new FoodSearchWidget component. - Replaced ProductSearchField and FoodSearchResultRow with integrated functionality in FoodSearchWidget. - Updated TreatmentsRootView to utilize the new FoodSearchWidget for searching and adding food items. - Enhanced ScannedProductList with improved layout and functionality for managing scanned products. - Removed unnecessary keyboard visibility state management from various views.
