+++ Trio/Sources/Modules/BarcodeScanner/View/BarcodeScannerRootView.swift
@@ -23,7 +23,6 @@ extension BarcodeScanner {
     struct RootView: BaseView {
         let resolver: Resolver
         var showListInitially: Bool = false
-        var onAddTreatments: ((Decimal, Decimal, Decimal, String) -> Void)?
 
         @ObservedObject var state: StateModel
         @State private var isEditingFromList = false
@@ -36,15 +35,11 @@ extension BarcodeScanner {
             resolver: Resolver,
             state: StateModel,
             showListInitially: Bool = false,
-            onAddTreatments: ((Decimal, Decimal, Decimal, String) -> Void)? = nil,
             onDismiss: (() -> Void)? = nil
         ) {
             self.resolver = resolver
             _state = ObservedObject(wrappedValue: state)
             self.showListInitially = showListInitially
-            self.onAddTreatments = onAddTreatments
-            // Wire optional callback into the state so it can call back when user selects "Add to Treatments"
-            self.state.onAddTreatments = onAddTreatments
             self.state.onDismiss = onDismiss
         }
 
@@ -223,14 +218,6 @@ extension BarcodeScanner {
                     }
                 }
             }
-            .onChange(of: focusedItemID) { _, newValue in
-                if newValue != nil {
-                    state.isKeyboardVisible = true
-                    state.isScanning = false
-                } else {
-                    state.isKeyboardVisible = false
-                }
-            }
         }
 
         // MARK: - Loading View
@@ -259,6 +246,7 @@ extension BarcodeScanner {
                     case .authorized:
                         CodeScannerView(
                             codeTypes: [.ean13, .ean8, .upce, .code128, .code39],
+                            scanMode: .continuous,
                             requiresPhotoOutput: false,
                             isTorchOn: state.isTorchOn,
                             isPaused: !state.isScanning,
@@ -271,7 +259,6 @@ extension BarcodeScanner {
                                 }
                             }
                         )
-                        .id(state.codeScannerViewID)
 
                     case .notDetermined:
                         VStack {
@@ -321,13 +308,16 @@ extension BarcodeScanner {
         private var listViewContent: some View {
             List {
                 Section {
-                    // Search Bar
-                    BarcodeScanner.ProductSearchField(
+                    // Food Search Widget
+                    BarcodeScanner.FoodSearchWidget(
                         searchText: $state.searchQuery,
                         isFocused: $isSearchFocused,
-                        onSubmit: {
-                            state.performFoodSearch()
-                        },
+                        isSearching: state.isSearching,
+                        searchError: state.searchError,
+                        searchResults: state.searchResults,
+                        hasMoreSearchResults: state.hasMoreSearchResults,
+                        isLoadingMoreSearchResults: state.isLoadingMoreSearchResults,
+                        onSubmit: { state.performFoodSearch() },
                         onClear: {
                             state.searchQuery = ""
                             state.searchResults = []
@@ -336,176 +326,37 @@ extension BarcodeScanner {
                             state.searchResults = []
                             state.searchError = nil
                             state.hasMoreSearchResults = false
-                        }
+                        },
+                        onItemAdd: { item in
+                            var mutableItem = item
+                            mutableItem.amount = item.servingQuantity ?? 100
+                            state.scannedProducts.append(mutableItem)
+                            state.searchQuery = ""
+                            state.searchResults = []
+                            isSearchFocused = false
+                        },
+                        onLoadMore: { state.loadMoreSearchResults() }
                     )
                     .listRowInsets(EdgeInsets(top: 0, leading: 0, bottom: 10, trailing: 0))
-
-                    // Search Results progress spinner and error display
-                    if state.isSearching {
-                        HStack {
-                            Spacer()
-                            ProgressView()
-                                .padding(.vertical, 8)
-                            Spacer()
-                        }
-                        .listRowInsets(EdgeInsets())
-                        .listRowBackground(Color.clear)
-                        .listRowSeparator(.hidden)
-                    } else if let error = state.searchError {
-                        Text(error)
-                            .font(.caption)
-                            .foregroundStyle(.red)
-                            .listRowInsets(EdgeInsets())
-                            .listRowBackground(Color.clear)
-                            .listRowSeparator(.hidden)
-                    } else if !state.searchResults.isEmpty {
-                        ForEach(state.searchResults) { item in
-                            BarcodeScanner.FoodSearchResultRow(item: item) {
-                                withAnimation {
-                                    var mutableItem = item
-                                    mutableItem.amount = item.servingQuantity ?? 100
-                                    state.scannedProducts.append(mutableItem)
-                                    state.searchQuery = ""
-                                    state.searchResults = []
-                                    isSearchFocused = false
-                                }
-                            }
-                            .listRowInsets(EdgeInsets(top: 4, leading: 0, bottom: 4, trailing: 0))
-                        }
-
-                        // Show "Show more results" button if there are more results
-                        if state.hasMoreSearchResults {
-                            Button {
-                                state.loadMoreSearchResults()
-                            } label: {
-                                HStack {
-                                    if state.isLoadingMoreSearchResults {
-                                        ProgressView()
-                                            .scaleEffect(0.9)
-                                    } else {
-                                        Text("Show 4 more results")
-                                            .font(.caption.weight(.medium))
-                                        Image(systemName: "chevron.down")
-                                            .font(.caption)
-                                    }
-                                }
-                                .foregroundStyle(.blue)
-                                .frame(maxWidth: .infinity)
-                                .padding(.vertical, 8)
-                            }
-                            .buttonStyle(.plain)
-                            .disabled(state.isLoadingMoreSearchResults)
-                        }
-                    }
-
-                    // Info about how much you scanned and total carbs
-                    if !state.scannedProducts.isEmpty {
-                        listHeader
-                            .listRowInsets(EdgeInsets(top: 4, leading: 0, bottom: 0, trailing: 0))
-                    }
-
-                    if state.scannedProducts.isEmpty, state.searchResults.isEmpty, !state.isSearching {
-                        emptyListView
-                    }
                 }
                 .listRowBackground(Color.clear)
                 .listRowSeparator(.hidden)
 
-                // Scanned products
-                // Seperate Section for better grouping and styling
-                if !state.scannedProducts.isEmpty {
-                    Section {
-                        ForEach(state.scannedProducts) { item in
-                            ScannedProductRow(item: item, state: state, focusedItemID: $focusedItemID)
-                                .listRowInsets(EdgeInsets())
-                                .contextMenu {
-                                    actionButtonsForScannedProduct(for: item)
-                                }
-                                .swipeActions(edge: .trailing, allowsFullSwipe: true) {
-                                    actionButtonsForScannedProduct(for: item)
-                                }
-                                .padding(15)
-                        }
+                BarcodeScanner.ScannedProductsListBody(
+                    state: state,
+                    focusedItemID: $focusedItemID,
+                    onEdit: { product in
+                        state.editScannedProduct(product)
+                        isEditingFromList = true
+                        state.isEditingFromList = true
+                        showEditorCard = true
                     }
-                    .listRowBackground(Color.chart)
-                }
+                )
             }
             .scrollContentBackground(.hidden)
             .listStyle(.insetGrouped)
             .listSectionSpacing(8)
             .contentMargins(.top, 4)
-        }
-
-        private func actionButtonsForScannedProduct(for product: BarcodeScanner.FoodItem) -> some View {
-            Group {
-                Button(role: .destructive) {
-                    withAnimation {
-                        state.removeScannedProduct(product)
-                    }
-                } label: {
-                    Label("Delete", systemImage: "trash.fill")
-                }
-                .tint(.red)
-
-                Button {
-                    state.editScannedProduct(product)
-                    isEditingFromList = true
-                    state.isEditingFromList = true
-                    showEditorCard = true
-                } label: {
-                    Label("Edit", systemImage: "pencil")
-                }
-                .tint(.blue)
-            }
-        }
-
-        private var emptyListView: some View {
-            VStack(spacing: 20) {
-                Spacer()
-                Image(systemName: "barcode.viewfinder")
-                    .font(.system(size: 60))
-                    .foregroundStyle(.secondary)
-                Text(String(localized: "No items yet"))
-                    .font(.title3.weight(.medium))
-                Text(String(localized: "Scan barcodes or search to add items."))
-                    .font(.subheadline)
-                    .foregroundStyle(.secondary)
-                    .multilineTextAlignment(.center)
-                    .padding(.horizontal)
-                Button {
-                    state.showListView = false
-                } label: {
-                    HStack {
-                        Image(systemName: "barcode.viewfinder")
-                        Text(String(localized: "Start Scanning"))
-                    }
-                    .font(.subheadline.weight(.semibold))
-                    .padding(.horizontal)
-                }
-                .buttonStyle(.borderedProminent)
-                .padding(.top, 8)
-                Spacer()
-            }
-            .frame(maxWidth: .infinity)
-            .padding(.vertical, 60)
-            .listRowBackground(Color.clear)
-            .listRowSeparator(.hidden)
-        }
-
-        private var listHeader: some View {
-            VStack(alignment: .leading, spacing: 8) {
-                Text(
-                    "\(state.scannedProducts.count) Item\(state.scannedProducts.count == 1 ? "" : "s")"
-                )
-                .font(.title2)
-                .bold()
-
-                HStack(spacing: 16) {
-                    Text("total \(state.scannedCarbs, specifier: "%.1f") g of carbs")
-                        .foregroundStyle(.blue)
-                }
-                .font(.subheadline)
-            }
         }
     }
 }
